Add bower to dev dependencies and remove unnecessary installs.
It's unnecessary (and in some cases undesirable) to install a global
gulp/bower.  The version of the global dependency may conflict between
multiple projects (one project could be built on gulp 3 and another on
gulp 2 for instance).  By encouraging use of the locally installed
gulp/bower, this reduces the chances of issues with people's
environments.  It also makes the setup steps a couple commands shorter
:).

I added an install-libs script to npm.  This makes it easy for users to
execute bower without having to learn anything about the `PATH`
environment variable or how to execute scripts in a hidden relative
directory (`node_modules/.bin`).  For the docker setup, however, I did
things a cleaner way by adding the full npm packages bin directory to
the docker container's path.  This makes it so that any commands
provided by the various npm packages can be easily executed.  For
example, `docker run -v $(pwd):/code pogo bower install`.