Add a docker-compose configuration for easier setup.
This provides a way of starting it with docker without using the long
`docker run` command.  Instead, if the user has docker-compose
installed, they can just run `docker-compose up` to start the service.

This also starts the container with a data volume for the
http-mitm-proxy certs so that they will be stored on the host in the
normal location (`.http-mitm-proxy`) and persist from one container to
the next, even across image updates.